+++ src/main/java/kr/hanjari/backend/domain/club/application/command/CodeGenerator.java
@@ -1,0 +1,6 @@
+package kr.hanjari.backend.domain.club.application.command;
+
+public interface CodeGenerator {
+
+    String reissueCode(Long clubId);
+}
