+++ src/main/java/kr/hanjari/backend/domain/club/application/command/impl/CodeGeneratorImpl.java
@@ -1,0 +1,50 @@
+package kr.hanjari.backend.domain.club.application.command.impl;
+
+import java.security.SecureRandom;
+import java.util.Random;
+import kr.hanjari.backend.domain.club.application.command.CodeGenerator;
+import kr.hanjari.backend.domain.club.domain.entity.Club;
+import kr.hanjari.backend.domain.club.domain.repository.ClubRepository;
+import kr.hanjari.backend.global.payload.code.status.ErrorStatus;
+import kr.hanjari.backend.global.payload.exception.GeneralException;
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Component;
+import org.springframework.transaction.annotation.Transactional;
+
+@Component
+@RequiredArgsConstructor
+@Transactional
+public class CodeGeneratorImpl implements CodeGenerator {
+    private final ClubRepository clubRepository;
+    private final Random random = new SecureRandom();
+
+    private static final String CHAR_POOL = "ABCDEFGHIJKLMNOPQRSTUVWXYZ0123456789"; // 코드에 사용할 문자 집합
+    private static final int CODE_LENGTH = 6; // 코드 길이
+
+    @Override
+    public String reissueCode(Long clubId) {
+        String code;
+        boolean isUnique;
+
+        do {
+            code = generateRandomCode();
+            isUnique = !clubRepository.existsByCode(code);
+        } while (!isUnique);
+
+        Club club = clubRepository.findById(clubId)
+                .orElseThrow(() -> new GeneralException(ErrorStatus._CLUB_NOT_FOUND));
+        club.updateCode(code);
+
+        return code;
+    }
+
+    private String generateRandomCode() {
+        StringBuilder codeBuilder = new StringBuilder(CODE_LENGTH);
+        for (int i = 0; i < CODE_LENGTH; i++) {
+            int index = random.nextInt(CHAR_POOL.length());
+            codeBuilder.append(CHAR_POOL.charAt(index));
+        }
+
+        return codeBuilder.toString();
+    }
+}
