+++ src/main/java/kr/hanjari/backend/domain/club/application/query/MailSender.java
@@ -1,0 +1,6 @@
+package kr.hanjari.backend.domain.club.application.query;
+
+public interface MailSender {
+
+    void sendEmail(String email, String clubName, String verificationCode, String loginUrl);
+}
